+++ package.json
@@ -32,6 +32,7 @@
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.36.0",
 		"flowbite-svelte": "^0.47.4",
+		"flowbite-svelte-blocks": "^1.1.4",
 		"flowbite-svelte-icons": "^2.0.2",
 		"globals": "^15.0.0",
 		"prettier": "^3.3.2",
