+++ src/lib/components/pages/auth/ForgotPassword.svelte
