+++ src/lib/components/pages/auth/ProfileLock.svelte
