+++ src/lib/components/ui/buttons/BackToHome.svelte
@@ -1,0 +1,18 @@
+<script lang="ts">
+	import { Button } from 'flowbite-svelte';
+</script>
+
+<Button href="/">
+	<svg
+		class="-ml-1 mr-2 h-5 w-5"
+		fill="currentColor"
+		viewBox="0 0 20 20"
+		xmlns="http://www.w3.org/2000/svg"
+		><path
+			fill-rule="evenodd"
+			d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
+			clip-rule="evenodd"
+		/></svg
+	>
+	Go back home
+</Button>
